Fix timer bug from PROCESS level to THREAD level
Previously, the timer measures the time duration with
"CLOCK_PROCESS_CPUTIME_ID" argument. It times throughout whole
process which means if the process spawns n threads. It will sum
up all of the time among these threads and it is not what we want.
Thus, switch the timer to thread level and ignore the slight
difference at the beginning of each thread.